requirements.txt
# ============================================================
#  AI Git Guard — Python Dependencies
# ============================================================

# GitHub API
httpx==0.27.0

# LLM providers
ollama==0.2.1
anthropic==0.28.0
openai==1.35.0

# Data models
pydantic==2.7.4
python-dotenv==1.0.1

# CLI interface
click==8.1.7
rich==13.7.1
prompt_toolkit==3.0.47

# Output formats
reportlab==4.2.2
openpyxl==3.1.4

# Caching
# sqlite3 is built-in to Python stdlib

# Testing
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
respx==0.22.0
